Moving method to Signature class
