+++ src/main/java/smile/identity/core/Utilities.java
@@ -184,7 +184,7 @@ public class Utilities {
             String signature = (String) responseJson.get(Signature.SIGNATURE_KEY);
             Signature sigObj = new Signature(partner_id, api_key);
             Boolean valid = false;
-            Boolean useSignature = useSignature(options);
+            Boolean useSignature = Signature.useSignature(options);
             
             if (useSignature) {
             	Long tstmpLng = new SimpleDateFormat(Signature.DATE_TIME_FORMAT).parse(timestamp).getTime();
@@ -206,7 +206,7 @@ public class Utilities {
         JSONObject body = new JSONObject();
         Boolean returnImages = (Boolean) options.get("return_images");
         Boolean returnHistory = (Boolean) options.get("return_history");
-        Boolean useSignature = useSignature(options);
+        Boolean useSignature = Signature.useSignature(options);
         
         Signature sigObj = new Signature(partner_id, api_key);
         body.put((useSignature) ? Signature.SIGNATURE_KEY : Signature.SEC_KEY, (useSignature) ? sigObj.getSignature(timestamp) : sigObj.getSecKey(timestamp));
@@ -251,12 +251,5 @@ public class Utilities {
         return HttpClientBuilder.create()
                 .setDefaultRequestConfig(config)
                 .build();
-    }
-
-    public static Boolean useSignature(JSONObject options){
-        if (options.containsKey(Signature.SIGNATURE_KEY)) {
-            return (Boolean) options.get(Signature.SIGNATURE_KEY);
-        }
-        return true;
     }
 }
